final version of our develop! should be ready to be merged into master branch

+++ examinationsprojekt1/src/global_variables.c
@@ -11,10 +11,10 @@
 
 uint16_t user_setpoint = 0;
 double user_P = 0;
-uint16_t user_I = 0;
+double user_I = 0;
 double user_D = 0;
 
-uint16_t tank1_value = 0;
-uint16_t tank2_value = 0;
-int32_t error_value = 0;
+double tank1_value = 0;
+double tank2_value = 0;
+double error_value = 0;
 uint32_t output_value = 0;
